requirements.txt
requests
python-dotenv
google-cloud-bigquery
google-auth
